% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{check_valid_structure}
\alias{check_valid_structure}
\title{Check that a list-like object has valid structure for}
\usage{
check_valid_structure(l)
}
\description{
Check that a list-like object has valid structure for
}
\keyword{internal}
