% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{cumrisk-weight}
\alias{cumrisk-weight}
\alias{product_limit}
\alias{product_limit,v_rcensored-method}
\title{Weighting functions}
\usage{
product_limit(x, ...)

\S4method{product_limit}{v_rcensored}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{additonal arguments passed to weighting functions where needed}
}
\value{
a \code{numeric} vector of weights
}
\description{
These functions are designed to estimate the probability of being \emph{uncensored}.
}
\section{Functions}{
\itemize{
\item \code{product_limit}: used product limit (Kaplan-Meier) estimator

\item \code{product_limit,v_rcensored-method}: used product limit (Kaplan-Meier) estimator
}}

\section{Product-limit}{


When censoring and outcomes occur at the same time, the \code{cumrisk} function may
give incorrect results when weighting by product-limit survival estimates.
To handle this, the \code{product_limit} function add a small amount (\code{1e-11}) to
any censors at tied times. Uncensoring probabilities for \code{infinite} times are
replaced with the \code{min} of the estimated probabilities; that is, the
probability from the last time is carried forward.
}

