% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{cumrisk}
\alias{cumrisk}
\alias{cumrisk,list,missing,maybeTime-method}
\alias{cumrisk,data.frame,numeric,maybeTime-method}
\title{Estimate cumulative risk}
\usage{
cumrisk(x, w, times = NULL)

\S4method{cumrisk}{list,missing,maybeTime}(x, w, times = NULL)

\S4method{cumrisk}{data.frame,numeric,maybeTime}(x, w, times = NULL)
}
\arguments{
\item{x}{the object from which to estimate risk. See details.}

\item{w}{a vector of weights}

\item{times}{a numeric vector of times at which to estimate risk. Defaults to
the times at which events occurred.}
}
\description{
Estimate cumulative risk
}
\details{
TODO
}
