% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{gather_cumrisk_data}
\alias{gather_cumrisk_data}
\title{Gathers data necessary in cumrisk estimator from a single variable list
element.}
\usage{
gather_cumrisk_data(vs)
}
\arguments{
\item{vs}{a named \code{list} containing data for the \code{cumrisk} estimator}
}
\description{
Gathers data necessary in cumrisk estimator from a single variable list
element.
}
\keyword{internal}
