% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{make_cumrisk_estimator}
\alias{make_cumrisk_estimator}
\title{An estimator of cumulative risk}
\usage{
make_cumrisk_estimator(vsl)
}
\arguments{
\item{vsl}{a \code{list} of \code{list}s where each element should contain \code{Y} (the
event time), \code{C} (the censoring time), and \code{PrDel} (the estimated
probability of being uncensored)}
}
\description{
An estimator of cumulative risk
}
