% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{validate_cumrisk_df}
\alias{validate_cumrisk_df}
\title{Check that a list structure is ready for cumrisk}
\usage{
validate_cumrisk_df(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{list}}
}
\description{
Check that a list structure is ready for cumrisk
}
\keyword{internal}
