% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{cumrisk-weight}
\alias{cumrisk-weight}
\alias{product_limit}
\alias{product_limit,v_rcensored-method}
\title{Weighting functions}
\usage{
product_limit(x)

\S4method{product_limit}{v_rcensored}(x)
}
\arguments{
\item{x}{the object}
}
\value{
a \code{numeric} vector of weights
}
\description{
These functions are designed to estimate the probability of being \emph{uncensored}.
}
\section{Functions}{
\itemize{
\item \code{product_limit}: used product limit (Kaplan-Meier) estimator

\item \code{product_limit,v_rcensored-method}: used product limit (Kaplan-Meier) estimator
}}

