% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{cumrisk}
\alias{cumrisk}
\alias{cumrisk,list,missing,maybeTime-method}
\alias{cumrisk,data.frame,numeric,maybeTime-method}
\alias{cumrisk,data.frame,function,maybeTime-method}
\alias{cumrisk,v_rcensored,numeric,maybeTime-method}
\alias{cumrisk,v_rcensored,function,maybeTime-method}
\title{Estimate cumulative risk}
\usage{
cumrisk(x, w, times = NULL, ...)

\S4method{cumrisk}{list,missing,maybeTime}(x, w, times = NULL, ...)

\S4method{cumrisk}{data.frame,numeric,maybeTime}(x, w, times = NULL, ...)

\S4method{cumrisk}{data.frame,`function`,maybeTime}(x, w, times = NULL, ...)

\S4method{cumrisk}{v_rcensored,numeric,maybeTime}(x, w, times = NULL, ...)

\S4method{cumrisk}{v_rcensored,`function`,maybeTime}(x, w, times = NULL, ...)
}
\arguments{
\item{x}{the object from which to estimate risk. See details.}

\item{w}{a vector of weights \emph{or} a \code{function} used to estimate weights (see
details). See \code{\link{cumrisk-weight}} for functions
provided by the \code{riskimator} package.}

\item{times}{a numeric vector of times at which to estimate risk. Defaults to
the times at which events occurred.}

\item{...}{\emph{If} \code{w} is a \code{function}, additional arguments are passed to that
function. Otherwise, additional arguments are ignored.}
}
\value{
a \code{list} of:
\itemize{
\item \code{time}: a vector a times at which cumulative risk was evaluated
\item \code{estimate}: the estimated cumulative risk for each time
}
}
\description{
Estimates cumulative risk of right-censored outcomes using the estimator
described in Hubbard, Van der Laan, and Robins (2000):
\deqn{
\widehat{\Pr}(Y < t) = \frac{1}{n} \sum_{i = 1}^n \frac{\Delta_i I(Y_i < t)} { \widehat{\Pr}(\Delta_i = 1)}
}
where \eqn{\Delta_i = I(Y_i < C_i)}, \eqn{Y_i} is the time to the event of
interest, and \eqn{C_i} is the time to censoring.
}
\details{
\code{cumrisk} accepts three types of inputs: a \code{list}, a \code{data.frame},
or a \code{\link[stype]{v_rcensored}} vector. \code{list} inputs should be a
transposed \code{data.frame} with one element per observation, each element
being a \code{list} containing either:
\itemize{
\item \code{Y}, \code{C}, and \code{PrDel} (the time to outcome, time to censoring,
and probablity of being uncensored)
\item \code{Ymin}, \code{Del}, and \code{PrDel} (\code{min(Y, C)},  an indicator of being uncensored,
and probablity of being uncensored)
}

\code{list}s are expected to sorted by \code{min(Y, C)}; this is not validated.

The more user-friendly input formats are \code{data.frame} and \code{v_rcensored}. With
both these types, a vector of \code{numeric} weights is required. See
\code{\link{cumrisk-weight}} for functions provided by the
\code{riskimator} package. \code{data.frame} inputs must contain either of the variable
set described above. An error is thrown if the data is not sorted by
\code{min(Y, C)}. With \code{v_rcensored} vectors sorting is handled by \code{cumrisk}.

By default, estimates of cumulative risk at event times are returned. The
\code{times} argument can be used to estimate times at other points.
}
\section{Using weight \code{function}s}{


The \code{w}eight argument can be function of the form \code{function(x, ...){}} that
returns a \code{numeric} vector, where \code{x} matches the type passed to \code{x} in
\code{cumrisk}. For example, if you pass a \code{data.frame} to \code{cumrisk}, the \code{w}
function should take a \code{data.frame}.
}

\references{
Hubbard A, Laan M van der, Robins JM. Nonparametric locally
efficient estimation of the treatment specific survival distribution with
right censored data and covariates in observational studies.
In: Halloran M, Berry D, editors. Statistical models in epidemiology,
the environment, and clinical trials. New York, NY: Springer; 2000.
}
