% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumrisk.R
\name{make_cumrisk_estimator}
\alias{make_cumrisk_estimator}
\title{Creates an estimator for cumulative risk}
\usage{
make_cumrisk_estimator(vsl)
}
\arguments{
\item{vsl}{a \code{list} of \code{list}s where each element should contain \code{Y} (the
event time), \code{C} (the censoring time), and \code{PrDel} (the estimated
probability of being uncensored)}
}
\description{
Creates the following estimator:
\deqn{
\widehat{\Pr}(Y < t) = \frac{1}{n} \sum_{i = 1}^n \frac{\Delta_i I(Y_i < t)} { \widehat{\Pr}(\Delta_i = 1)}
}
where \eqn{\Delta_i = I(Y_i < C_i)}, \eqn{Y_i} is the time to the event of
interest, and \eqn{C_i} is the time to censoring.
}
\keyword{internal}
